src/requirements.txt
lxml
python-docx
Flask==2.0.0
Flask-WTF
Flask-Bootstrap4
